github_upload/requirements.txt
numpy==1.24.4
opencv_python==4.9.0.80
opencv_python_headless==4.10.0.84
pandas==2.0.3
Pillow==11.1.0
roboflow==1.1.51
supervision==0.25.1
torch==2.2.0
transformers==4.46.3
ultralytics==8.3.67